requirements.txt
mysql-connector-python
numpy
matplotlib
python-dotenv